フロントエンド：ヘッダーファイル分割
