+++ apps/frontend/src/utils/constants.ts
@@ -1,0 +1,4 @@
+// toC向け
+export const CUSTOMER_NAVIGATION_LINKS = [
+  { label: "ソロ活をシェアする", href: "/reviews" },
+];
